Initial lack of seriousness
We're being flexible with the use of "lack of" here mostly because we, and by we I mean I, seriously don't expect this to work. But I do expect it to provide ample opportunities for awful jokes and poor attempts at humor while coaxing coding assistants into ridiculous attempts at creating a deliverable app.